To avoid confusion, split tests which need to run isolated because of cross-test pollution in a separate module.
