+++ tox.ini
@@ -36,9 +36,9 @@ commands =
     # isolated from other tests
     # NOTE 2: We write junit XML file for nicer visualization on Circle CI
     rm -rf test-results
-    py.test tests/unit/ -vv --durations=5 -m "not memory_leak and not json_lib" --junitxml=test-results/junit-1.xml
+    py.test tests/unit/ -vv --durations=5 -m "not memory_leak and not json_lib_reload" --junitxml=test-results/junit-1.xml
     py.test tests/unit/test_memory_leaks.py -vv --durations=5 --junitxml=test-results/junit-2.xml
-    py.test tests/unit/util/json_util_test.py -vv --durations=5 -m "json_lib" --junitxml=test-results/junit-3.xml
+    py.test tests/unit/util/default_json_lib.py -vv --durations=5 -m "json_lib_reload" --junitxml=test-results/junit-3.xml
 
 # Target which generates documentation for all the monitors
 [testenv:generate-monitor-docs]
